Update `pre-commit` config for `isort`
When run under `pre-commit`, `isort` is unable to detect `docent` as a
first-party library, so specify it explicitly as an argument in
`pyproject.toml`.

Also explicitly specify the top-level `pyproject.toml` in the
`pre-commit` config, since otherwise `isort` may pick up the one in
`cli/`, if the first file passed to it is in that subdirectory.

Also update to the latest version while we're here.