Add initial prototype of job batching
Early stab at moving some heavy LLM computations out of the HTTP
nendpoint handler and into a separate worker.

No UI changes per se, but now you can start a search, close Docent, and
come back to it, and results will have continued being computed in the
background.

The `docent worker` command spins up a background worker process.

Incoming search queries are routed to a queue in Redis; a worker pops
each query from there, runs the API requests to fulfill the query, and
sends the results back on a unique new queue (keyed by the
just-generated job ID). Workers handle saving results to Postgres, which
continues to be the store for long-term data.

Uses `arq`, a wrapper around Redis primitives that provides a nice job
queue interface. I'm now not sure how much we need that yet for this
scenario -- for returning progress from the workers back to the
listener, I tried out just using two basic Redis commands, which might
be all we need here.

To think about: managing the lifecycle of the channels used for
returning results. Right now, there is a queue for each job ID; if
everything is read out of the queue, it gets cleaned up automatically by
Redis, but if the connection is lost, I think the results will just sit
there forever. (Redis key TTL is a thing, maybe use that.)

To think about: how to organize more complicated workflows. Currently,
the queue name is hardcoded and there's just one path for processing
this one type of job.