Upload results continuously during search
Previously, search results were all being uploaded to the database in
one batch at the end of the search. That is useful to reduce load on the
database, but that shouldn't be much of a concern until farther down the
road, and uploading results as we go means that progress is more
visible. If/when we get to a point where load really is an issue, we can
switch this over to batching once a second or something.